searchStaffs 関連の処理を Staff.php ではなく、StaffPDFCreater.php に任せるようにした。(Issue #41)
