+++ app/Models/StaffPDFCreater.php
@@ -4,6 +4,8 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
+use DB;
+
 use ZendPdf\PdfDocument;
 use ZendPdf\Font;
 use ZendPdf\Resource\Extractor;
@@ -110,4 +112,29 @@ class StaffPDFCreater extends Model {
         header ('Content-Disposition:', 'inline;');
         echo $pdfDocument->render();
     }
+
+    /*
+     * searchStaffs テーブルから特定のイベントに紐付いたデータを取得する。
+     */
+    public function getSearch($id) {
+        $searchStaffs = DB::table('searchStaffs')
+            ->where('event_id', $id)
+            ->get();
+
+        $getSearch = end($searchStaffs);
+
+        return $getSearch;
+    }
+
+    /*
+     * 検索結果から PDF を出力するために staffs テーブルから検索結果を返す
+     */
+    public function searchStaffs($id, $content, $text) {
+        $searchStaff = Staff::where('event_id', $id)
+            ->where($content, 'like', '%'. $text .'%')
+            ->get();
+
+        return $searchStaff;
+    }
+
 }
